- change expression analytics type to float - close proxied netcdfs
git-svn-id: svn://bar.ebi.ac.uk/branches/atlas-standalone@11547 2913f559-6b04-0410-9a09-c530ee9f5186

+++ atlas-dao/src/main/java/uk/ac/ebi/gxa/dao/AtlasDAO.java
@@ -1275,8 +1275,8 @@ public class AtlasDAO {
                                          String property,
                                          String propertyValue,
                                          int[] designElements,
-                                         double[] pValues,
-                                         double[] tStatistics) {
+                                         float[] pValues,
+                                         float[] tStatistics) {
         // execute this procedure...
         /*
         PROCEDURE A2_AnalyticsSet(
@@ -1828,8 +1828,8 @@ public class AtlasDAO {
     }
 
     private SqlTypeValue convertExpressionAnalyticsToOracleARRAY(final int[] designElements,
-                                                                 final double[] pValues,
-                                                                 final double[] tStatistics) {
+                                                                 final float[] pValues,
+                                                                 final float[] tStatistics) {
         if (designElements == null || pValues == null || tStatistics == null ||
                 designElements.length != pValues.length || pValues.length != tStatistics.length) {
             throw new RuntimeException(
